Fix: follow 추가 기능 수정
follow 추가 기능 수정


Former-commit-id: e29302f6cf974701531867500b4ee070fdc5ed9f

+++ user/src/main/java/com/wanyviny/user/domain/follow/service/FollowServiceImpl.java
@@ -61,17 +61,18 @@ public class FollowServiceImpl implements FollowService {
         User following = userRepository.findById(followingId).orElseThrow(
                 () -> new IllegalArgumentException("followingId에 해당하는 유저가 없습니다.")
         );
+        if (!followRepository.existsFollowByUserId_IdAndFollowingId_Id(userId, followingId)) {
+            followRepository.save(Follow.builder()
+                    .userId(user)
+                    .followingId(following)
+                    .build());
 
-        followRepository.save(Follow.builder()
-                .userId(user)
-                .followingId(following)
-                .build());
+            String title = "FOLLOW!";
+            String body = user.getNickname() + "님이 당신을 팔로우했습니다.";
+            String token = following.getPhoneId();
 
-        String title = "FOLLOW!";
-        String body = user.getNickname() + "님이 당신을 팔로우했습니다.";
-        String token = following.getPhoneId();
-
-        firebasePushAlarm(title, body, token);
+            firebasePushAlarm(title, body, token);
+        }
     }
 
     @Override
@@ -83,8 +84,8 @@ public class FollowServiceImpl implements FollowService {
         User following = userRepository.findById(followingId).orElseThrow(
                 () -> new IllegalArgumentException("followingId에 해당하는 유저가 없습니다.")
         );
-
-        followRepository.deleteByUserIdAndFollowingId(user, following);
+        if (followRepository.existsFollowByUserId_IdAndFollowingId_Id(userId, followingId))
+            followRepository.deleteByUserIdAndFollowingId(user, following);
     }
 
     @Transactional
